rough add ExpressionManipulator
