+++ src/ast/CMakeLists.txt
@@ -1,4 +1,5 @@
 SET(ast_SOURCES
   ExpressionAnalyzer.cpp
+  ExpressionManipulator.cpp
 )
 ADD_LIBRARY(ast STATIC ${ast_SOURCES})
